Python/kivy_mobile_app/requirements.txt
appdirs==1.4.3
certifi==2019.11.28
chardet==3.0.4
distlib==0.3.0
docutils==0.16
filelock==3.0.12
idna==2.9
importlib-metadata==1.5.0
Kivy==1.11.1
kivy-deps.angle==0.1.10
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.18
kivy-deps.sdl2==0.1.23
Kivy-Garden==0.1.4
mysql-connector==2.2.9
Pygments==2.5.2
pypiwin32==223
pywin32==227
requests==2.23.0
six==1.14.0
urllib3==1.25.8
virtualenv==20.0.7
zipp==3.1.0
